+++ lib/jazzy/source_declaration/access_control_level.rb
@@ -5,26 +5,52 @@ module Jazzy
 
       attr_reader :level
 
-      def initialize(declaration_string)
-        if declaration_string =~ /private\ /
+      ACCESSIBILITY_PRIVATE  = 'source.lang.swift.accessibility.private'
+      ACCESSIBILITY_INTERNAL = 'source.lang.swift.accessibility.internal'
+      ACCESSIBILITY_PUBLIC   = 'source.lang.swift.accessibility.public'
+
+      def initialize(accessibility)
+        if accessibility == ACCESSIBILITY_PRIVATE
           @level = :private
-        elsif declaration_string =~ /public\ /
-          @level = :public
-        else
+        elsif accessibility == ACCESSIBILITY_INTERNAL
           @level = :internal
+        elsif accessibility == ACCESSIBILITY_PUBLIC
+          @level = :public
+        end
+      end
+
+      def self.from_doc(doc)
+        accessibility = doc['key.accessibility']
+        if accessibility
+          acl = new(accessibility)
+          if acl
+            return acl
+          end
+        end
+        acl = from_explicit_declaration(doc['key.parsed_declaration'])
+        acl || AccessControlLevel.public # fallback on public ACL
+      end
+
+      def self.from_explicit_declaration(declaration_string)
+        if declaration_string =~ /private\ /
+          return AccessControlLevel.private
+        elsif declaration_string =~ /public\ /
+          return AccessControlLevel.public
+        elsif declaration_string =~ /internal\ /
+          return AccessControlLevel.internal
         end
       end
 
       def self.private
-        AccessControlLevel.new('private ')
+        new(ACCESSIBILITY_PRIVATE)
       end
 
       def self.internal
-        AccessControlLevel.new('internal ')
+        new(ACCESSIBILITY_INTERNAL)
       end
 
       def self.public
-        AccessControlLevel.new('public ')
+        new(ACCESSIBILITY_PUBLIC)
       end
 
       LEVELS = {