+++ lib/jazzy/sourcekitten.rb
@@ -56,7 +56,7 @@ module Jazzy
       expected_xcode_select_path =
         Pathname('/Applications/Xcode.app/Contents/Developer')
       return if xcode_developer_directory == expected_xcode_select_path
-      raise 'Please install or symlink Xcode 6.1 or 6.2 in ' \
+      raise 'Please install or symlink Xcode 6.3 in ' \
             "#{expected_xcode_select_path} and set as active developer " \
             'directory by running `sudo xcode-select -s ' \
             "#{expected_xcode_select_path}`"
@@ -70,7 +70,7 @@ module Jazzy
     def self.assert_swift_version
       swift_version = `xcrun swift --version` =~ /Swift version ([\d\.]+)/ &&
         Regexp.last_match[1]
-      expected_swift_version = '1.1'
+      expected_swift_version = '1.2'
       return if swift_version == expected_swift_version
       raise "Jazzy only works with Swift #{expected_swift_version}."
     end
@@ -104,8 +104,7 @@ module Jazzy
       # Always document extensions, since we can't tell what ACL they are
       return true if doc['key.kind'] == 'source.lang.swift.decl.extension'
 
-      SourceDeclaration::AccessControlLevel.new(doc['key.annotated_decl']) >=
-        @min_acl
+      SourceDeclaration::AccessControlLevel.from_doc(doc) >= @min_acl
     end
 
     def self.process_undocumented_token(doc, declaration)
@@ -212,7 +211,7 @@ module Jazzy
         declaration.name = doc['key.name']
         declaration.mark = current_mark
         acl = SourceDeclaration::AccessControlLevel.new(
-          doc['key.annotated_decl'],
+          doc['key.accessibility'],
         )
         declaration.access_control_level = acl
         declaration.start_line = doc['key.parsed_scope.start']